feat(#17): app config; testes de integração atendente
